+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [2.1.10]
+
+### Changed
+
+* The command [workflows](README.md#workflows) shows enabled workflows
+* the workflow parameter allows version specification: ex. `taxa:1.1.0`
+
+
 ## [2.1.9]
 
 ### Changed